### 2021-03-23 #3132 @whedon
[PRE REVIEW]: Assessment, Testing and Comparison of Statistical Models using R
**Submitting author:** @strengejacke (<a href="http://orcid.org/0000-0002-8895-3206">Daniel Lüdecke</a>)
**Repository:** <a href="https://github.com/easystats/performance" target ="_blank">https://github.com/easystats/performance</a>
**Version:** 0.7.1
