+++ NAMESPACE
@@ -526,6 +526,7 @@ importFrom(stats,fitted)
 importFrom(stats,fligner.test)
 importFrom(stats,formula)
 importFrom(stats,glm)
+importFrom(stats,influence)
 importFrom(stats,lm)
 importFrom(stats,mad)
 importFrom(stats,mahalanobis)
