### 2021-03-24 9e25cff
Update check_outliers.R
