### 2021-03-24 0b80a2f
Update check_model_diagnostics.R
