+++ R/check_model_diagnostics.R
@@ -184,7 +184,7 @@
     stringsAsFactors = FALSE
   )
   plot_data$Index <- 1:nrow(plot_data)
-  plot_data$Influential <- NA
+  plot_data$Influential <- "OK"
   plot_data$Influential[abs(plot_data$Cooks_Distance) >= max(cook_levels)] <- "Influential"
 
   attr(plot_data, "cook_levels") <- cook_levels
