### 2021-03-24 dfda49d
update citation and README with it
