### 2021-03-25 6dd076c
Update order of elements in title to match the order of the section
