### 2021-03-25 #3132 @Kevin-Mattheus-Moerman
@strengejacke thanks for your submission. I wanted to make a recommendation with regards to your paper title. It is more usual for JOSS paper titles to hint at the name of the package in the title, e.g. perhaps you can consider something like: `Performance:  an R Package for Assessment, Testing and Comparison of Statistical Models`.
