### 2021-03-25 easystats/performance#240 @mikldk
JOSS
Hi,

From openjournals/joss-reviews#3132 I saw your README and this example:

```r
model <- lm(mpg ~ wt + cyl, data = mtcars)
r2(model)
#> # R2 for Linear Regression
#>
#>        R2: 0.830
#>   adj. R2: 0.819

model <- glm(am ~ wt + cyl, data = mtcars, family = binomial)
r2(model)
#> $R2_Tjur
#> Tjur's R2
#>    0.7051

library(MASS)
data(housing)
model <- polr(Sat ~ Infl + Type + Cont, weights = Freq, data = housing)
r2(model)
#> $R2_Nagelkerke
#> Nagelkerke's R2
#>          0.1084
```

I was slightly confused that the `lm` version is "pretty printed", whereas the `glm`/`polr` seems more like a printed vector. Maybe I am misunderstanding something, or else it may be worth trying to make (well-known) models "pretty print" like `lm`.
