### 2021-03-25 easystats/performance#241 @mikldk
JOSS: Method details
In the JOSS paper you give this example:

```r
data(cars)
model<-lm(dist~speed,data =cars)
check_heteroscedasticity(model)
#> Warning: Heteroscedasticity (non-constant error variance) detected (p = 0.031).
```
What statistical test is used? Is it the Breush-Pagan test? I think you need to give details, potentially in the output or at least in the documentation of the function. The above only shows `check_heteroscedasticity()` but you need to verify this for all such functions.

I see that you have one implementation (`check_heteroscedasticity.default()`). Is it correct that this will give the correct results for all model types? Also e.g. Poisson regression? Or is it only `lm` models?

Ping openjournals/joss-reviews#3132
