+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "performance: an R Package for Assessment, Testing and Comparison of Statistical Models."
+title: "performance: an R Package for Assessment, Testing and Comparison of Statistical Models"
 authors:
 - affiliation: 1
   name: Daniel Lüdecke
