+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: 23eb05ceceda1c44573b254dd8b96e92cd91f825
+pkgdown_sha: fe799d33217ac4b9ffc1df52b08233c6f68e85c7
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-03-25T12:10Z
+last_built: 2021-03-25T13:56Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
