+++ NAMESPACE
@@ -211,6 +211,7 @@ S3method(print,r2_generic)
 S3method(print,r2_mlm)
 S3method(print,r2_nakagawa)
 S3method(print,r2_nakagawa_by_group)
+S3method(print,r2_pseudo)
 S3method(print,test_likelihoodratio)
 S3method(print,test_performance)
 S3method(print_html,compare_performance)
@@ -270,6 +271,7 @@ S3method(r2,mixed)
 S3method(r2,mlm)
 S3method(r2,mlogit)
 S3method(r2,mmclogit)
+S3method(r2,model_fit)
 S3method(r2,multinom)
 S3method(r2,negbinirr)
 S3method(r2,negbinmfx)

[…]
