+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: 9a5a8f5cd29bb06616c1452d1b9b5cab7cebf7c2
+pkgdown_sha: 672b8fc2d10d98ed9bab242c2af56662e5983e9b
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-03-26T10:19Z
+last_built: 2021-03-27T00:58Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
