### 2021-03-27 easystats/performance#241 @mikldk
Are you sure the test is valid for all the possible models for which your code provides a hypothesis test?
