+++ R/check_heteroscedasticity.R
@@ -12,9 +12,7 @@
 #' @note There is also a \href{https://easystats.github.io/see/articles/performance.html}{\code{plot()}-method} implemented in the \href{https://easystats.github.io/see/}{\pkg{see}-package}.
 #'
 #' @details This test of the hypothesis of (non-)constant error is also called
-#' \emph{Breusch-Pagan test} (\cite{1979}). \code{check_heteroscedasticity()}
-#' should return correct results for all models for which Pearson residuals can
-#' be calculated, and for which residual degrees of freedom are available.
+#' \emph{Breusch-Pagan test} (\cite{1979}).
 #'
 #' @references Breusch, T. S., and Pagan, A. R. (1979) A simple test for heteroscedasticity and random coefficient variation. Econometrica 47, 1287–1294.
 #'
@@ -37,6 +35,11 @@ check_heteroscedasticity <- function(x, ...) {
 
 #' @export
 check_heteroscedasticity.default <- function(x, ...) {
+  # only for linear models
+  if (!insight::model_info(x)$is_linear) {
+    stop("This Breusch-Pagan Test currently only works Gaussian models.")
+  }
+
   r <- .pearson_residuals(x)
   S.sq <- insight::get_df(x, type = "residual") * .sigma(x)^2 / sum(!is.na(r))
 
