### 2021-03-27 easystats/performance#241 @strengejacke
I specified the valid input. It is not limited to lm per se, but to Gaussian models, which is checked now. Some packages, like speedlm, should also work, despite probably not being of class "lm".
