### 2021-03-28 easystats/performance#241 @mikldk
Note that a linear model is not necessarily Gaussian (at least theoretically), but I'm not sure how you check it and how it impacts the test. Thanks for including the details 👍.
