### 2021-03-29 e12b4d6
support model_fit
