+++ R/model_performance.lm.R
@@ -312,6 +312,9 @@ model_performance.betaor <- model_performance.logitor
 #' @export
 model_performance.betamfx <- model_performance.logitor
 
+#' @export
+model_performance.model_fit <- model_performance.logitor
+
 
 
 
