+++ NAMESPACE
@@ -27,6 +27,7 @@ S3method(check_distribution,numeric)
 S3method(check_heteroscedasticity,default)
 S3method(check_homogeneity,default)
 S3method(check_model,default)
+S3method(check_model,model_fit)
 S3method(check_normality,default)
 S3method(check_normality,glmmTMB)
 S3method(check_normality,merMod)
@@ -37,11 +38,13 @@ S3method(check_outliers,glmrob)
 S3method(check_outliers,lme)
 S3method(check_outliers,lmrob)
 S3method(check_outliers,numeric)
+S3method(check_overdispersion,default)
 S3method(check_overdispersion,fixest)
 S3method(check_overdispersion,glm)
 S3method(check_overdispersion,glmmTMB)
 S3method(check_overdispersion,glmx)
 S3method(check_overdispersion,merMod)
+S3method(check_overdispersion,model_fit)
 S3method(check_overdispersion,negbin)
 S3method(check_overdispersion,negbinirr)
 S3method(check_overdispersion,negbinmfx)
@@ -108,6 +111,7 @@ S3method(model_performance,merMod)
 S3method(model_performance,mixed)
 S3method(model_performance,mixor)
 S3method(model_performance,mlogit)
+S3method(model_performance,model_fit)
 S3method(model_performance,multinom)
 S3method(model_performance,negbinirr)
 S3method(model_performance,negbinmfx)
