### 2021-03-29 66b8cfa
Add Tolerance to VIF printing
