+++ R/print-methods.R
@@ -582,9 +582,11 @@ print.check_collinearity <- function(x, ...) {
 
 .print_collinearity <- function(x) {
   vifs <- x$VIF
+  x$Tolerance <- 1/x$VIF
 
   x$VIF <- sprintf("%.2f", x$VIF)
   x$SE_factor <- sprintf("%.2f", x$SE_factor)
+  x$Tolerance <- sprintf("%.2f", x$Tolerance)
 
   colnames(x)[3] <- "Increased SE"
 
