### 2021-03-29 #3139 @whedon
[REVIEW]: performance: An R Package for Assessment, Comparison1and Testing of Statistical Models
**Submitting author:** @strengejacke (<a href="http://orcid.org/0000-0002-8895-3206">Daniel Lüdecke</a>)
**Repository:** <a href="https://github.com/easystats/performance" target ="_blank">https://github.com/easystats/performance</a>
**Version:** 0.7.1
