### 2021-03-29 #3139 @mikldk
@martinju, @gjo11 (aided by @sachsmc): Thanks for agreeing to review. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If possible create issues (and cross-reference) in the submission's repository to avoid too specific discussions in this review thread.

@strengejacke Please note the error "Failed to discover a Statement of need section in paper". Please fix, recompile and notify the reviewers.
