### 2021-03-30 a7bebd4
update figure, check_model mor customizable
