### 2021-03-30 f0597a5
Update NEWS.md
