### 2021-03-31 #3139 @gjo11
Thank you. This seems like a really nice and useful package!
Just a few things: The output from the paper does not reflect what is produced by the current version of the package (the version under review is 0.7.1). I suppose they may have been generated using an older version.

1.
The following example is given in the paper:

library(see)
model <- lm(mpg ~ wt + am + gear + vs * cyl, data = mtcars)
check_model(model)

Running the last command, the package wants to use "qqplotr", which should likely be listed as "suggests".

Also, the visuals are slightly different from those in the paper, which may have been generated using an older version.

2.
Running the example

library(brms)
set.seed(123)
model <- brm(mpg ~ wt + (1 | cyl) + (1 + wt | gear), data = mtcars)
icc(model)

outputs

Adjusted ICC: 0.929
Conditional ICC: 0.797

which is similar although not exactly the same as in the paper.

3.
The example

test_performance(lm1, lm2, lm3, lm4)

outputs

Name | Model |     BF
---------------------
lm1  |    lm |
lm2  |    lm | > 1000
lm3  |    lm | > 1000
lm4  |    lm | > 1000
Each model is compared to lm1.

rather than the output in the paper.

4.
The indices of the output of the example

test_bf(lm1, lm2, lm3, lm4)

are not prefixed by "lm" as in the output in the paper.
