### 2021-03-31 f4e9569
improve plot display sizes in README
