+++ README.md
@@ -5,7 +5,7 @@
 [![downloads](http://cranlogs.r-pkg.org/badges/performance)](https://cran.r-project.org/package=performance)
 [![total](https://cranlogs.r-pkg.org/badges/grand-total/performance)](https://cranlogs.r-pkg.org/)
 
-***Test if your model is a good model\!***
+***Test if your model is a good model!***
 
 A crucial aspect when building regression models is to evaluate the
 quality of modelfit. It is important to investigate how well models fit
@@ -80,7 +80,7 @@ There is a nice introduction into the package on
 
 ## The *performance* workflow
 
-![](man/figures/figure_workflow.png)
+<img src="man/figures/figure_workflow.png" width="75%" />
 
 ### Assessing model quality
 
@@ -131,8 +131,15 @@ slopes or nested random effects (Johnson 2014; Nakagawa, Johnson, and
 Schielzeth 2017).
 
 ``` r
+set.seed(123)
 library(rstanarm)
-model <- stan_glmer(Petal.Length ~ Petal.Width + (1 | Species), data = iris, cores = 4)
+
+model <- stan_glmer(
+  Petal.Length ~ Petal.Width + (1 | Species),
+  data = iris,
+  cores = 4
+)
+
 r2(model)
 #> # Bayesian R2 with Standard Error
 #> 
@@ -250,12 +257,16 @@ set.seed(123)
 sleepstudy$mygrp <- sample(1:5, size = 180, replace = TRUE)
 sleepstudy$mysubgrp <- NA
 for (i in 1:5) {
-    filter_group <- sleepstudy$mygrp == i
-    sleepstudy$mysubgrp[filter_group] <- sample(1:30, size = sum(filter_group), replace = TRUE)
+  filter_group <- sleepstudy$mygrp == i
+  sleepstudy$mysubgrp[filter_group] <-
+    sample(1:30, size = sum(filter_group), replace = TRUE)
 }
 
 # fit strange model
-model <- lmer(Reaction ~ Days + (1 | mygrp/mysubgrp) + (1 | Subject), data = sleepstudy)
+model <- lmer(
+  Reaction ~ Days + (1 | mygrp / mysubgrp) + (1 | Subject),
+  data = sleepstudy
+)
 
 check_singularity(model)
 #> [1] TRUE
@@ -287,11 +298,14 @@ check_heteroscedasticity(model)
 `check_model()`.
 
 ``` r
-model <- lm(mpg ~ wt * cyl + gear, data = mtcars)
+# defining a model
+model <- lm(mpg ~ wt + am + gear + vs * cyl, data = mtcars)
+
+# checking model assumptions
 check_model(model)
 ```
 
-![](man/figures/unnamed-chunk-15-1.png)<!-- -->
+<img src="man/figures/unnamed-chunk-15-1.png" width="100%" />
 
 ### Model performance summaries
 
@@ -386,7 +400,7 @@ installed).
 plot(compare_performance(m1, m2, m4, rank = TRUE))
 ```
 
-![](man/figures/unnamed-chunk-21-1.png)<!-- -->
+<img src="man/figures/unnamed-chunk-21-1.png" width="100%" />
 
 ### Testing models
 
@@ -401,7 +415,8 @@ data(iris)
 lm1 <- lm(Sepal.Length ~ Species, data = iris)
 lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
 lm3 <- lm(Sepal.Length ~ Species * Sepal.Width, data = iris)
-lm4 <- lm(Sepal.Length ~ Species * Sepal.Width + Petal.Length + Petal.Width, data = iris)
+lm4 <- lm(Sepal.Length ~ Species * Sepal.Width + 
[…]
