+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: 38f0d8a2db6e26027bc4da8ded7d035e4ddaf987
+pkgdown_sha: 5a62000c8d09754e3ea22788e302b1230537b766
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-03-30T09:47Z
+last_built: 2021-03-31T18:01Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
