+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: 20c4390517f26189bf7743ae530cb0ec3f8e8800
+pkgdown_sha: a48e6280894bcfd718e04deaabccd2d84fefb48d
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-01T00:37Z
+last_built: 2021-04-01T05:15Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
