### 2021-04-01 02bd207
Update test_bf.R
https://github.com/vincentarelbundock/performance/commit/b04362cb3a0a5b5f36e2a36c8f3ee0c096e54dce
