+++ R/test_bf.R
@@ -51,7 +51,11 @@ test_bf.ListModels <- function(objects, reference = 1, ...) {
   if (reference == "sequential") {
     # TODO: Double check that formula and whether it works for increasing and
     # decreasing order.
-    rez$BF <- rez$BF / c(1, rez$BF[1:nrow(rez) - 1])
+
+    # For increasing
+    rez$BF <- exp(c(NA, diff(log(rez$BF))))
+    # For decreasing
+    # rez$BF <- exp(c(-diff(log(rez$BF)), NA))
   }
 
   rez$BF[ref] <- NA
