+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: a48e6280894bcfd718e04deaabccd2d84fefb48d
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-01T08:06Z
+last_built: 2021-04-01T14:29Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
