### 2021-04-01 d4a5877
Update NEWS.md
