### 2021-04-01 #3139 @strengejacke
@gjo11 Thanks for your comments. I addressed your issues, however, some are related to dependencies from GitHub versions of packages.

1. The easystats-packages are designed with developers in mind, that may use the functions from packages like _performance_ in their packages. Thus, we try to keep dependencies as low as possible. Thus, all plotting functions are located in the _see_ package (which is suggested by _performance_), and the _see_ package - in turn - suggests qqplotr when necessary.

The second point you mentioned, the slightly different figure in the paper, might be due to the fact that we have largely revised the plot-style from `check_model()` due to several user feedback (including improvements for color-blinds, see https://github.com/easystats/see/issues/97). So installing _performance_ and _see_ from GitHub should give you the same results.

2. Not sure if this was due to changes in RStan, brms or insight? I have checked the example on two different computers (both Windows, R 4.0.4, and both yielded the same result), and updated the numbers in the paper.

3. & 4. I updated the paper, these were indeed outputs from "older" versions of the `print()` function.
