### 2021-04-02 7e9e2d4
multiline display of code in docs
https://easystats.github.io/performance/reference/check_outliers.html#threshold-specification
