+++ R/check_outliers.R
@@ -17,7 +17,7 @@
 #'   considered as outlier can be recovered with the \code{as.data.frame}
 #'   function.
 #'
-#' @note There is also a \href{https://easystats.github.io/see/articles/performance.html}{\code{plot()}-method} implemented in the \href{https://easystats.github.io/see/}{\pkg{see}-package}. \strong{Please note} that the range of the distance-values along the y-axis is re-scaled to range from 0 to 1!
+#' @note There is also a \href{https://easystats.github.io/see/articles/performance.html}{\code{plot()}-method} implemented in the \href{https://easystats.github.io/see/}{\pkg{see}-package}. \strong{Please note} that the range of the distance-values along the y-axis is re-scaled to range from 0 to 1.
 #'
 #' @details Outliers can be defined as particularly influential observations. Most methods rely on the computation of some distance metric, and the observations greater than a certain threshold are considered outliers. Importantly, outliers detection methods are meant to provide information to the researcher, rather than being an automatized procedure which mindless application is a substitute for thinking.
 #'
@@ -108,7 +108,7 @@
 #' \subsection{Threshold specification}{
 #' Default thresholds are currently specified as follows:
 #'
-#' \code{
+#' \preformatted{
 #' list(
 #'   zscore = stats::qnorm(p = 1 - 0.025),
 #'   iqr = 1.5,
