### 2021-04-03 0f9be99
allow optional args (...) to be passed to check_outliers_mahala #252
close #252
