+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: a48e6280894bcfd718e04deaabccd2d84fefb48d
+pkgdown_sha: 5b5baea274400b242e0a2194fe9d392f753d3508
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-02T05:06Z
+last_built: 2021-04-03T01:47Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
