+++ man/performance_aicc.Rd
@@ -3,7 +3,7 @@
 \name{performance_aicc}
 \alias{performance_aicc}
 \alias{performance_aic}
-\title{Compute AIC and second order AIC}
+\title{Compute the AIC or second-order AIC}
 \usage{
 performance_aicc(x, ...)
 
@@ -18,11 +18,10 @@ performance_aic(x, ...)
 Numeric, the AIC or AICc value.
 }
 \description{
-Compute the second-order Akaike's information criterion (AICc).
-The second-order (or small sample) is a AIC with a correction for small sample
-sizes. \code{performance_aic()} is a small wrapper that returns the AIC. It is
+Compute the AIC or the second-order Akaike's information criterion (AICc).
+\code{performance_aic()} is a small wrapper that returns the AIC. It is
 a generic function that also works for some models that don't have a AIC method
-(like Tweedie models).
+(like Tweedie models). \code{performance_aicc()} returns the second-order (or "small sample") AIC that incorporates a correction for small sample sizes.
 }
 \examples{
 m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
