+++ NAMESPACE
@@ -524,6 +524,7 @@ importFrom(stats,cor)
 importFrom(stats,cov)
 importFrom(stats,cov2cor)
 importFrom(stats,dbinom)
+importFrom(stats,dchisq)
 importFrom(stats,density)
 importFrom(stats,df.residual)
 importFrom(stats,dnbinom)
@@ -534,6 +535,7 @@ importFrom(stats,fligner.test)
 importFrom(stats,formula)
 importFrom(stats,glm)
 importFrom(stats,influence)
+importFrom(stats,integrate)
 importFrom(stats,lm)
 importFrom(stats,mad)
 importFrom(stats,mahalanobis)
@@ -565,6 +567,7 @@ importFrom(stats,shapiro.test)
 importFrom(stats,simulate)
 importFrom(stats,summary.lm)
 importFrom(stats,terms)
+importFrom(stats,uniroot)
 importFrom(stats,update)
 importFrom(stats,var)
 importFrom(stats,vcov)
