### 2021-04-05 166ea15
Update confint_r2.R
