### 2021-04-05 d8c0034
Update confint_r2.R
