+++ R/confint_r2.R
@@ -43,13 +43,13 @@ dRsq <- function(K1, R2_pop, R2_obs, p, nobs) {
   NCP <- R2_pop / (1 - R2_pop)
   F1_obs <- ((nobs - p - 1) / p) * (R2_obs / (1 - R2_obs))
   exp(log(
-    stats::pf(
+    suppressWarnings(stats::pf(
       q = F1_obs,
       df1 = p,
       df2 = (nobs - p - 1),
       ncp = NCP * K1,
       lower.tail = FALSE
-    )
+    ))
   ) + stats::dchisq(x = K1, df = (nobs - 1), log = TRUE))
 }
 
