+++ R/r2.R
@@ -6,6 +6,10 @@
 #'   R2, pseudo-R2, or marginal / adjusted R2 values are returned.
 #'
 #' @param model A statistical model.
+#' @param ci Confidence Interval (CI) level. Default is \code{NULL}. Confidence
+#'   intervals for R2 can be calculated based on different methods, see
+#'   \code{ci_method}.
+#' @param ci_method to do
 #' @param ... Arguments passed down to the related r2-methods.
 #' @inheritParams r2_nakagawa
 #'
@@ -50,7 +54,12 @@ r2 <- function(model, ...) {
 
 #' @importFrom insight print_color get_response get_predicted
 #' @export
-r2.default <- function(model, verbose = TRUE, ...) {
+r2.default <- function(model, ci = NULL, ci_method = "analytical", verbose = TRUE, ...) {
+  ci_method <- match.arg(ci_method, choices = c("analytical", "bootstrap"))
+
+  # check input
+  ci <- .check_r2_ci_args(ci, ci_method, "bootstrap", verbose)
+
   out <- tryCatch(
     {
       if (insight::model_info(model)$is_binomial) {
@@ -724,4 +733,22 @@ r2.DirichletRegModel <- function(model, ...) {
   names(out$R2_Nagelkerke) <- "Nagelkerke's R2"
   class(out) <- c("r2_pseudo", class(out))
   out
+}
+
+
+
+
+
+# helper -------------------
+
+.check_r2_ci_args <- function(ci = NULL, ci_method = "bootstrap", valid_ci_method = NULL, verbose = TRUE) {
+  if (!is.null(ci) && !is.na(ci)) {
+    if (!is.null(valid_ci_method) && !ci_method %in% valid_ci_method) {
+      if (verbose) {
+        warning(paste0("Method '", ci_method, "' to compute confidence intervals for R2 not supported."), call. = FALSE)
+      }
+      return(NULL)
+    }
+  }
+  ci
 }
