### 2021-04-06 f2ff1ef
rename confint_r2 -> r2_ci for easystats consistentcy
