+++ man/r2.Rd
@@ -2,10 +2,13 @@
 % Please edit documentation in R/r2.R
 \name{r2}
 \alias{r2}
+\alias{r2.default}
 \alias{r2.merMod}
 \title{Compute the model's R2}
 \usage{
 r2(model, ...)
+
+\method{r2}{default}(model, ci = NULL, ci_method = "analytical", verbose = TRUE, ...)
 
 \method{r2}{merMod}(model, tolerance = 1e-05, ...)
 }
@@ -14,18 +17,18 @@ r2(model, ...)
 
 \item{...}{Arguments passed down to the related r2-methods.}
 
+\item{ci}{Confidence Interval (CI) level. Default is \code{NULL}. Confidence
+intervals for R2 can be calculated based on different methods, see
+\code{ci_method}.}
+
+\item{ci_method}{to do}
+
 \item{tolerance}{Tolerance for singularity check of random effects, to decide
 whether to compute random effect variances for the conditional r-squared
 or not. Indicates up to which value the convergence result is accepted. When
 \code{r2_nakagawa()} returns a warning, stating that random effect variances
 can't be computed (and thus, the conditional r-squared is \code{NA}),
 decrease the tolerance-level. See also \code{\link{check_singularity}}.}
-
-\item{ci}{Confidence Interval (CI) level. Default is \code{NULL}. Confidence
-intervals for R2 can be calculated based on different methods, see
-\code{ci_method}.}
-
-\item{ci_method}{to do}
 }
 \value{
 Returns a list containing values related to the most appropriate R2
