+++ R/helpers.R
@@ -165,11 +165,18 @@
 
 
 .remove_backticks_from_parameter_names <- function(x) {
-  if (is.data.frame(x) && "Parameter" %in% colnames(x)) {
-    x$Parameter <- gsub("`", "", x$Parameter, fixed = TRUE)
+  if (is.data.frame(x)) {
+    if ("Parameter" %in% colnames(x)) {
+      x$Parameter <- gsub("`", "", x$Parameter, fixed = TRUE)
+    }
+    if ("Term" %in% colnames(x)) {
+      x$Term <- gsub("`", "", x$Term, fixed = TRUE)
+    }
   }
+
   x
 }
+
 
 
 #' @importFrom insight get_sigma
