+++ DESCRIPTION
@@ -49,8 +49,8 @@ BugReports: https://github.com/easystats/performance/issues
 Depends:
     R (>= 3.4)
 Imports:
-    bayestestR (>= 0.7.5),
-    insight (>= 0.11.0),
+    bayestestR (>= 0.8.2),
+    insight (>= 0.13.2),
     stats,
     utils
 Suggests:
