+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ddafb4ac3eff2ec4e5988b6b13bc4585e9df186c
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-06T08:24Z
+last_built: 2021-04-06T08:30Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
