### 2021-04-07 5409ae5
fix and improve examples of check_outliers
