+++ R/check_outliers.R
@@ -68,7 +68,7 @@
 #' \item \strong{Robust Mahalanobis Distance}:
 #' A robust version of Mahalanobis distance using an Orthogonalized
 #' Gnanadesikan-Kettenring pairwise estimator (Gnanadesikan \& Kettenring, 1972).
-#' Requires the \pkg{bigutilsr} package.
+#' Requires the \pkg{bigutilsr} package. See the \code{bigutilsr::dist_ogk()} function.
 #'
 #' \item \strong{Minimum Covariance Determinant (MCD)}:
 #' Another robust version of Mahalanobis. Leys et al. (2018) argue that Mahalanobis Distance is not a robust way to
@@ -137,10 +137,37 @@
 #' }
 #'
 #' @examples
-#' # Univariate
-#' check_outliers(mtcars$mpg)
+#' data <- mtcars  # Size nrow(data) = 32
 #'
-#' # Multivariate
+#' # For single variables ------------------------------------------------------
+#' outliers_list <- check_outliers(data$mpg)  # Find outliers
+#' outliers_list  # Show the row index of the outliers
+#' as.numeric(outliers_list)  # The object is a binary vector...
+#' filtered_data <- data[!outliers_list, ] # And can be used to filter a dataframe
+#' nrow(filtered_data)  # New size, 28 (4 outliers removed)
+#'
+#'
+#' # For dataframes ------------------------------------------------------------
+#' check_outliers(data)  # It works the same way on dataframes
+#'
+#' # You can also use multiple methods at once
+#' outliers_list <- check_outliers(data, method = c("mahalanobis",
+#'                                                  "iqr",
+#'                                                  "zscore"))
+#' outliers_list
+#' # Using `as.data.frame()`, we can access more details!
+#' outliers_info <- as.data.frame(outliers_list)
+#' head(outliers_info)
+#' outliers_info$Outlier  # Including the probability of being an outlier
+#' # And we can be more stringent in our outliers removal process
+#' filtered_data <- data[outliers_info$Outlier < 0.1, ]
+#'
+#' \dontrun{
+#' # You can also run all the methods
+#' check_outliers(data, method = "all")
+#' }
+#'
+#' # For statistical models ----------------------------------------------------
 #' # select only mpg and disp (continuous)
 #' mt1 <- mtcars[, c(1, 3, 4)]
 #' # create some fake outliers and attach outliers to main df
@@ -148,21 +175,16 @@
 #' # fit model with outliers
 #' model <- lm(disp ~ mpg + hp, data = mt2)
 #'
-#' ol <- check_outliers(model)
-#' # plot(ol)
-#' insight::get_data(model)[ol, ]
+#' outliers_list <- check_outliers(model)
+#' # plot(outliers_list)
+#' insight::get_data(model)[outliers_list, ]  # Show outliers data
 #'
 #' if (require("MASS")) {
 #'   check_outliers(model, method = c("mahalabonis", "mcd"))
 #' }
-#' \dontrun{
 #' # This one takes some seconds to finish...
 #' check_outliers(model, method = "ics")
 #'
-#' # For dataframes
-#' check_outliers(mtcars)
-#' check_outliers(mtcars, method = "all")
-#' }
 #' @importFrom insight n_obs get_predictors get_data
 #' @importFrom stats cooks.distance mahalanobis cov
 #' @export
@@ -531,7 +553,8 @@ as.numeric.check_outliers <- function(x, ...) {
 
 
   # Compute
-  out$Distance_Robust <- bigutilsr::dist_ogk(as.matrix(x))
+  U <- svd(scale(x))$u
+  out$Distance_Robust <- bigutilsr::dist_ogk(U)
 
   # Filter
   out$Outlier_Robust <- as.numeric(out$Distance_Robust > threshold)
