+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: ddafb4ac3eff2ec4e5988b6b13bc4585e9df186c
+pkgdown_sha: 7a5834b0177a78acd65e168e55674104e0931d61
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-06T08:30Z
+last_built: 2021-04-07T14:47Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
