+++ DESCRIPTION
@@ -49,7 +49,7 @@ BugReports: https://github.com/easystats/performance/issues
 Depends:
     R (>= 3.4)
 Imports:
-    bayestestR (>= 0.8.2),
+    bayestestR (>= 0.9.0),
     insight (>= 0.13.2),
     stats,
     utils
@@ -104,5 +104,5 @@ Suggests:
     VGAM
 Encoding: UTF-8
 Language: en-US
-RoxygenNote: 7.1.1
+RoxygenNote: 7.1.1.9001
 Config/testthat/edition: 3
