+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: 7a5834b0177a78acd65e168e55674104e0931d61
+pkgdown_sha: e2f601eee4bfbddd168816eec6ddb70dfb6a72eb
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-07T15:40Z
+last_built: 2021-04-09T07:21Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
