+++ R/check_normality.R
@@ -47,8 +47,15 @@ check_normality <- function(x, ...) {
 }
 
 
+#' @importFrom insight model_info
 #' @export
 check_normality.default <- function(x, ...) {
+  # valid model?
+  if (!insight::model_info(x)$is_linear) {
+    message("Checking normality of residuals is only useful an appropriate assumption for linear models.")
+    return(NULL)
+  }
+
   # check for normality of residuals
   p.val <- .check_normality(stats::rstandard(x), x)
 
@@ -72,6 +79,12 @@ check_normality.merMod <- function(x, effects = c("fixed", "random"), ...) {
   # args
   effects <- match.arg(effects)
   info <- insight::model_info(x)
+
+  # valid model?
+  if (!info$is_linear) {
+    message("Checking normality of residuals is only useful an appropriate assumption for linear models.")
+    return(NULL)
+  }
 
   if (effects == "random") {
     if (!requireNamespace("lme4", quietly = TRUE)) {
