+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: performance
 Title: Assessment of Regression Models Performance
-Version: 0.7.1
+Version: 0.7.1.1
 Authors@R: 
     c(person(given = "Daniel",
              family = "Lüdecke",
@@ -104,5 +104,5 @@ Suggests:
     VGAM
 Encoding: UTF-8
 Language: en-US
-RoxygenNote: 7.1.1.9001
+RoxygenNote: 7.1.1
 Config/testthat/edition: 3
