+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: performance
 Title: Assessment of Regression Models Performance
-Version: 0.7.1.1
+Version: 0.7.1
 Authors@R: 
     c(person(given = "Daniel",
              family = "Lüdecke",
