### 2021-04-09 98d4374
Update NEWS.md
