+++ NEWS.md
@@ -1,4 +1,4 @@
-# performance 0.7.2
+# performance 0.7.1
 
 ## New supported models
 
