### 2021-04-09 #3139 @strengejacke
@gjo11 All relevant update to packages that were only available on GitHub are now accepted on CRAN, thus you should be able to reproduce all outputs from the paper with both the CRAN and GitHub versions. I re-generated the PDF proof with the changes you suggested, so the most recent draft was just generated by whedon.
