+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: e2f601eee4bfbddd168816eec6ddb70dfb6a72eb
+pkgdown_sha: 3dd1a1ad7740977491241c89eee15d9b58e75866
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-09T09:45Z
+last_built: 2021-04-09T16:01Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
