+++ NAMESPACE
@@ -24,10 +24,12 @@ S3method(check_convergence,default)
 S3method(check_convergence,merMod)
 S3method(check_distribution,default)
 S3method(check_distribution,numeric)
+S3method(check_heteroscedasticity,afex_aov)
 S3method(check_heteroscedasticity,default)
 S3method(check_homogeneity,default)
 S3method(check_model,default)
 S3method(check_model,model_fit)
+S3method(check_normality,afex_aov)
 S3method(check_normality,default)
 S3method(check_normality,glmmTMB)
 S3method(check_normality,merMod)
