### 2021-04-13 #3139 @martinju
Done with the review!

I agree with @gjo11 that performance seems like a very nice package. There is clearly a comprehensive amout of work underneath here. The package serves an important purpose, seems to be well programmed, is easy to use, well documented and tested and provides visually appealing plots and console print out. I only have a few comments. I am just putting them in a task list below here as I don't think it necessary to make seperate GitHub issues for these comments.

Author list
- [ ] I kindly ask you to briefly comment on the software contributions of author Philip Waggoner.
The commit history for the repository suggests this authors contributions are marginal. However, the repository seems to have been quite large already at first commit in the commit history.
Has this author contributed significantly to the software before the software reached git/GitHub, or has he contributed to the software in other ways?

Feature requests (Not required for acceptance):
- [ ] I really like the function check_model, which gives all appropriate visual checks for the model class. It would be very useful with a  non-visual counterpart which runs all appropriate test and provides p-values etc for them. It seems you don't have that implemented. It shouldn't be too hard based on the "manual" tests already implemented, or are there problems with this approach?

- [ ] All the *_performance functions computes in-sample based measures. From my point of view the word "performance" suggests how well the model works
out-of-sample. Would it be possible to add out-of-sample measures like RMSE, MAE, out-of-sample R^2, and for classification: AUC, precision, recall +++

Paper:

- [ ] Line 33-36: I don't quite like the last part of this sentence. Please write out
(generalized) linear MODEL, mixed-effects MODEL, and Bayesian MODEL or rephrase to improve clarity.

- [ ] Line 40: I don't understand the start of this sentence. Is there a typo in one of the ands or a missing commas?

- [ ] Line 47: Suggests overview -> demonstrate/take a look at/provide examples for

- [ ] Line 60: Delete "and more"? These are examples (ref "such as"), so of course there are more.

@strengejacke please respond to each of these comments.
