+++ R/performance_mse.R
@@ -104,3 +104,6 @@ performance_mse.betaor <- performance_mse.logitor
 
 #' @export
 performance_mse.betamfx <- performance_mse.logitor
+
+#' @export
+performance_mse.model_fit <- performance_mse.logitor
