+++ NAMESPACE
@@ -156,11 +156,23 @@ S3method(performance_logloss,default)
 S3method(performance_logloss,logitmfx)
 S3method(performance_logloss,logitor)
 S3method(performance_logloss,probitmfx)
+S3method(performance_mae,betamfx)
+S3method(performance_mae,betaor)
+S3method(performance_mae,default)
+S3method(performance_mae,logitmfx)
+S3method(performance_mae,logitor)
+S3method(performance_mae,model_fit)
+S3method(performance_mae,negbinirr)
+S3method(performance_mae,negbinmfx)
+S3method(performance_mae,poissonirr)
+S3method(performance_mae,poissonmfx)
+S3method(performance_mae,probitmfx)
 S3method(performance_mse,betamfx)
 S3method(performance_mse,betaor)
 S3method(performance_mse,default)
 S3method(performance_mse,logitmfx)
 S3method(performance_mse,logitor)
+S3method(performance_mse,model_fit)
 S3method(performance_mse,negbinirr)
 S3method(performance_mse,negbinmfx)
 S3method(performance_mse,poissonirr)
@@ -416,6 +428,7 @@ export(item_intercor)
 export(item_reliability)
 export(item_split_half)
 export(looic)
+export(mae)
 export(model_performance)
 export(mse)
 export(multicollinearity)
@@ -426,6 +439,7 @@ export(performance_aicc)
 export(performance_hosmer)
 export(performance_logloss)
 export(performance_lrt)
+export(performance_mae)
 export(performance_mse)
 export(performance_pcp)
 export(performance_rmse)
