### 2021-04-13 easystats/performance#266 @strengejacke
"text equivalent" to `check_model()`
> I really like the function check_model, which gives all appropriate visual checks for the model class. It would be very useful with a non-visual counterpart which runs all appropriate test and provides p-values etc for them. It seems you don't have that implemented. It shouldn't be too hard based on the "manual" tests already implemented, or are there problems with this approach?

(https://github.com/openjournals/joss-reviews/issues/3139#issuecomment-819005992)
