+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: 3dd1a1ad7740977491241c89eee15d9b58e75866
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-13T20:58Z
+last_built: 2021-04-13T21:04Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
