### 2021-04-13 #3139 @strengejacke
@martinju Thanks for your comments and suggestions. Let me summarize what I have done so far to address your issues.

- Author list: Philip is an early member of the team, who hasn't contributed much _code_ lately. However, he made substantial contributions to this paper. Thus, since he both contributed to package code, ideas and the documentation/paper, from our point of view, he definitely deserves an authorship role for this paper.

- Feature requests: Thanks for those suggestions! I have implemented `performance_mae()` (or its alias `mae()`) now as a first step. RMSE is already possible to calculate (`performance_rmse()` resp. `rmse()`). AUC is available in `bayestestR::area_under_curve()`, while the performance-package provides a function to calculate the ROC curve (`performance_roc()`). We also opened a new issue (https://github.com/easystats/performance/issues/266) resp. added your suggestions to an existing, related issue (https://github.com/easystats/performance/issues/36) in order to work on these ideas.

Paper:
- I have added the word "models" each, to improve clarity.
- I added a comma and removed parts of the sentence that were rather confusing. The paragraph now reads: _Beyond validity and quality checks, performance also includes plotting functions via the see package._
- I replaced "overview": The sentence now reads: _To demonstrate the efficiency of the package, we provide examples for a few functions, followed by a broader function that runs a comprehensive suite of checks in a single call._
- "read more" was deleted.

I hope I have addresses all issues to your satisfaction!
