### 2021-04-13 easystats/performance#266 @bwiernik
Personally, I think most of the tests and their p values are pretty much worthless. I'd strongly recommend most folks stick with the graphical measures, and I'm kind of okay with getting statistical tests requiring more hoops.
