### 2021-04-14 #3139 @martinju
Just a minor thing about this
> I have implemented `performance_mae()` (or its alias `mae()`) now as a first step. RMSE is already possible to calculate (`performance_rmse()` resp. `rmse()`)

My point here was to compute these measures for a model, but on out-of-sample data. From what I can see, the performance_* only does this for the data used to fit the model. Maybe you could just add two new optional arguments to the functions ytest and xtest which are used to compute the measures if provided?

If necessary we can continue the discussion in the relevant issue.
Dealing with this is not necessary to warrant JOSS publication.
