### 2021-04-14 #3139 @martinju
I am happy with the response and adjustments from @strengejacke to my comments.
(Note: I consider the above feature request comments that to be a feature for a potential future version and not something required at this stage.)

All points in the checklist are now checked off, and I recommend the current version of this package and paper for publication! Nice work @strengejacke et al.
