### 2021-04-15 06fe04a
Use model.matrix() in check_outliers instead of get_predictors()
ncol(data) needs to refer to the number for terms in the model function. The current approach is incorrect because it (1) doesn't separate factor variables into levels and (2) doesn't include the intercept.
