+++ R/check_outliers.R
@@ -184,8 +184,7 @@ check_outliers.default <- function(x, method = c("cook", "pareto"), threshold = 
   method <- match.arg(method, c("zscore", "iqr", "cook", "pareto", "mahalanobis", "robust", "mcd", "ics", "optics", "iforest", "lof"), several.ok = TRUE)
 
   # Remove non-numerics
-  data <- insight::get_predictors(x)
-  data <- data[, sapply(data, is.numeric), drop = FALSE]
+  data <- model.matrix(x)
 
 
 
