+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(as.data.frame,r2_bayes)
 S3method(as.numeric,check_outliers)
 S3method(check_autocorrelation,default)
 S3method(check_collinearity,MixMod)
+S3method(check_collinearity,afex_aov)
 S3method(check_collinearity,betamfx)
 S3method(check_collinearity,betaor)
 S3method(check_collinearity,default)
