### 2021-04-17 33cfe8a
add check_sphericity
