+++ NAMESPACE
@@ -62,6 +62,8 @@ S3method(check_singularity,glmmadmb)
 S3method(check_singularity,lme)
 S3method(check_singularity,merMod)
 S3method(check_singularity,rlmerMod)
+S3method(check_sphericity,afex_aov)
+S3method(check_sphericity,default)
 S3method(cronbachs_alpha,data.frame)
 S3method(cronbachs_alpha,matrix)
 S3method(cronbachs_alpha,parameters_pca)
@@ -409,6 +411,7 @@ export(check_outliers)
 export(check_overdispersion)
 export(check_posterior_predictions)
 export(check_singularity)
+export(check_sphericity)
 export(check_zeroinflation)
 export(compare_performance)
 export(cronbachs_alpha)
