### 2021-04-17 8019f2f
Update check_outliers.R
