### 2021-04-17 f2ca8a3
Update check_outliers.R
