+++ R/check_outliers.R
@@ -163,8 +163,6 @@
 #' check_outliers(mtcars)
 #' check_outliers(mtcars, method = "all")
 #' }
-#' @importFrom insight n_obs get_predictors get_data
-#' @importFrom stats cooks.distance mahalanobis cov
 #' @export
 check_outliers <- function(x, ...) {
   UseMethod("check_outliers")
@@ -391,7 +389,6 @@ as.numeric.check_outliers <- function(x, ...) {
 
 
 
-#' @importFrom stats median sd qnorm mad
 .check_outliers_zscore <- function(x, threshold = stats::qnorm(p = 1 - 0.025), robust = TRUE, method = "max") {
   # Standardize
   if (robust == FALSE) {
@@ -416,7 +413,6 @@ as.numeric.check_outliers <- function(x, ...) {
 
 
 
-#' @importFrom stats IQR quantile
 .check_outliers_iqr <- function(x, threshold = 1.5, method = "tukey") {
   d <- data.frame(Obs = 1:nrow(as.data.frame(x)))
   for (col in 1:ncol(as.data.frame(x))) {
@@ -455,7 +451,6 @@ as.numeric.check_outliers <- function(x, ...) {
 
 
 
-#' @importFrom stats cooks.distance
 .check_outliers_cook <- function(x, threshold = NULL) {
   # Compute
   d <- unname(stats::cooks.distance(x))
@@ -501,7 +496,6 @@ as.numeric.check_outliers <- function(x, ...) {
 
 
 
-#' @importFrom stats mahalanobis cov
 .check_outliers_mahalanobis <- function(x, threshold = NULL, ...) {
   out <- data.frame(Obs = 1:nrow(x))
 
@@ -653,8 +647,6 @@ as.numeric.check_outliers <- function(x, ...) {
 }
 
 
-# @importFrom utils packageVersion
-# @importFrom stats median qnorm mad sd predict
 # .check_outliers_iforest <- function(x, threshold = 0.025) {
 #   out <- data.frame(Obs = 1:nrow(x))
 #
