+++ R/check_distribution.R
@@ -49,9 +49,6 @@
 #'   check_distribution(model)
 #'   plot(check_distribution(model))
 #' }
-#' @importFrom bayestestR map_estimate
-#' @importFrom stats IQR density predict sd mad residuals
-#' @importFrom insight get_response
 #' @export
 check_distribution <- function(model) {
   UseMethod("check_distribution")
