+++ R/performance_roc.R
@@ -44,8 +44,6 @@
 #' # if you have \code{see} package installed, you can also plot comparison of
 #' # ROC curves for different models
 #' # if (require("see")) plot(performance_roc(m1, m2, m3))
-#' @importFrom stats predict
-#' @importFrom insight find_response get_data is_model model_info
 #' @export
 performance_roc <- function(x, ..., predictions, new_data) {
   dots <- list(...)
