+++ R/performance_mse.R
@@ -18,7 +18,6 @@
 #' data(mtcars)
 #' m <- lm(mpg ~ hp + gear, data = mtcars)
 #' performance_mse(m)
-#' @importFrom insight get_residuals
 #' @export
 performance_mse <- function(model, ...) {
   UseMethod("performance_mse")
