+++ NAMESPACE
@@ -458,71 +458,6 @@ export(test_performance)
 export(test_vuong)
 export(test_wald)
 export(variance_decomposition)
-importFrom(bayestestR,area_under_curve)
-importFrom(bayestestR,bayesfactor_models)
-importFrom(bayestestR,ci)
-importFrom(bayestestR,hdi)
-importFrom(bayestestR,point_estimate)
-importFrom(bayestestR,weighted_posteriors)
 importFrom(insight,display)
-importFrom(insight,ellipsis_info)
-importFrom(insight,export_table)
-importFrom(insight,find_algorithm)
-importFrom(insight,find_parameters)
-importFrom(insight,find_predictors)
-importFrom(insight,find_random)
-importFrom(insight,find_random_slopes)
-importFrom(insight,find_response)
-importFrom(insight,format_p)
-importFrom(insight,get_data)
-importFrom(insight,get_df)
-importFrom(insight,get_loglikelihood)
-importFrom(insight,get_parameters)
-importFrom(insight,get_predicted)
-importFrom(insight,get_residuals)
-importFrom(insight,get_response)
-importFrom(insight,get_variance)
-importFrom(insight,is_model)
-importFrom(insight,is_multivariate)
-importFrom(insight,model_info)
-importFrom(insight,n_obs)
-importFrom(insight,n_parameters)
-importFrom(insight,null_model)
-importFrom(insight,print_color)
 importFrom(insight,print_html)
 importFrom(insight,print_md)
-importFrom(stats,as.formula)
-importFrom(stats,binomial)
-importFrom(stats,coef)
-importFrom(stats,cor)
-importFrom(stats,dbinom)
-importFrom(stats,df.residual)
-importFrom(stats,dnbinom)
-importFrom(stats,dpois)
-importFrom(stats,family)
-importFrom(stats,fitted)
-importFrom(stats,glm)
-importFrom(stats,mad)
-importFrom(stats,median)
-importFrom(stats,model.matrix)
-importFrom(stats,na.omit)
-importFrom(stats,pchisq)
-importFrom(stats,pf)
-importFrom(stats,pnbinom)
-importFrom(stats,ppois)
-importFrom(stats,predict)
-importFrom(stats,predict.glm)
-importFrom(stats,qchisq)
-importFrom(stats,qf)
-importFrom(stats,qnorm)
-importFrom(stats,quantile)
-importFrom(stats,residuals)
-importFrom(stats,sd)
-importFrom(stats,simulate)
-importFrom(stats,summary.lm)
-importFrom(stats,update)
-importFrom(stats,var)
-importFrom(stats,weights)
-importFrom(stats,xtabs)
-importFrom(utils,capture.output)
-importFrom(utils,packageVersion)
