+++ DESCRIPTION
@@ -106,4 +106,5 @@ Encoding: UTF-8
 Language: en-US
 LazyData: true
 RoxygenNote: 7.1.1
+Remotes: easystats/insight
 Config/testthat/edition: 3
