### 2021-04-18 8c201dd
check_sphericity for mlm
