+++ R/check_sphericity.R
@@ -4,7 +4,7 @@
 #' @description Check model for violation of sphericity
 #'
 #' @param x A model object.
-#' @param ... Currently not used.
+#' @param ... Arguments passed to \code{car::Anova}.
 #'
 #' @return Invisibly returns the p-values of the test statistics. A p-value < 0.05
 #' indicates a violation of sphericity.
@@ -16,37 +16,13 @@ check_sphericity <- function(x, ...) {
 
 #' @export
 check_sphericity.default <- function(x, ...) {
-  stop("Test not supported yet for object of class ", class(x))
+  stop("Test not supported yet for object of class ", class(x)[1])
 }
 
-# #' @export
-# check_sphericity.mlm <- function(x, ...) {
-#   if (!requireNamespace("car"))
-#     stop("car required for this function to work.")
-#
-#   test <- summary(car::Anova(x))$sphericity.tests
-#
-#   p.val <- test[,2]
-#   if (any(p.val < .05)) {
-#     pp <- p.val[p.val < .05]
-#     pp <- paste0("\n - ", names(pp), " (", insight::format_p(pp), ")", collapse = "")
-#     insight::print_color(sprintf("Warning: Sphericity violated for: %s.\n", pp), "red")
-#   } else {
-#     pp <- insight::format_p(min(p.val))
-#     pp <- sub("=", ">", pp)
-#     insight::print_color(sprintf("OK: Data seems to be spherical (%s).\n", pp), "green")
-#   }
-#   invisible(p.val)
-# }
-
 #' @export
-#' @importFrom insight print_color format_p
-check_sphericity.afex_aov <- function(x, ...) {
-  if (!requireNamespace("afex"))
-    stop("afex required for this function to work.")
-
-  test <- afex::test_sphericity(x)
-
+check_sphericity.Anova.mlm <- function(x, ...) {
+  S <- summary(x, multivariate = FALSE, univariate = FALSE)
+  test <- S$sphericity.tests
 
   p.val <- test[,2]
   if (any(p.val < .05)) {
@@ -60,3 +36,23 @@ check_sphericity.afex_aov <- function(x, ...) {
   }
   invisible(p.val)
 }
+
+#' @export
+#' @importFrom insight print_color format_p
+check_sphericity.afex_aov <- function(x, ...) {
+  if (length(attr(afex_aov, "within")) == 0) {
+    stop("Mauchly Test of Sphericity is only aplicable to ANOVAs with within-subjects factors.")
+  }
+
+  check_sphericity.Anova.mlm(x, ...)
+}
+
+
+#' @export
+check_sphericity.mlm <- function(x, ...) {
+  if (!requireNamespace("car"))
+    stop("car required for this function to work.")
+
+  check_sphericity.Anova.mlm(car::Anova(x, ...))
+}
+
