+++ NAMESPACE
@@ -62,8 +62,10 @@ S3method(check_singularity,glmmadmb)
 S3method(check_singularity,lme)
 S3method(check_singularity,merMod)
 S3method(check_singularity,rlmerMod)
+S3method(check_sphericity,Anova.mlm)
 S3method(check_sphericity,afex_aov)
 S3method(check_sphericity,default)
+S3method(check_sphericity,mlm)
 S3method(cronbachs_alpha,data.frame)
 S3method(cronbachs_alpha,matrix)
 S3method(cronbachs_alpha,parameters_pca)
