### 2021-04-18 8b15387
fix: don't skip sph test
