### 2021-04-18 d89b8df
afex_aov homogen-var method
