+++ R/check_heteroscedasticity.R
@@ -12,8 +12,7 @@
 #' @note There is also a \href{https://easystats.github.io/see/articles/performance.html}{\code{plot()}-method} implemented in the \href{https://easystats.github.io/see/}{\pkg{see}-package}.
 #'
 #' @details This test of the hypothesis of (non-)constant error is also called
-#'   \emph{Breusch-Pagan test} (\cite{1979}). For \code{afex_aov} Levene's test
-#'   is preformed.
+#'   \emph{Breusch-Pagan test} (\cite{1979}).
 #'
 #' @references Breusch, T. S., and Pagan, A. R. (1979) A simple test for heteroscedasticity and random coefficient variation. Econometrica 47, 1287–1294.
 #'
@@ -74,36 +73,6 @@ check_heteroscedasticity.default <- function(x, ...) {
   class(p.val) <- unique(c("check_heteroscedasticity", "see_check_heteroscedasticity", class(p.val)))
 
   invisible(p.val)
-}
-
-
-#' @export
-check_heteroscedasticity.afex_aov <- function(x, ...) {
-  if (!requireNamespace("afex"))
-    stop("afex required for this function to work.")
-
-  test <- afex::test_levene(x)
-
-  is_covar <- sapply(attr(x,'between'), is.null)
-  if (any(is_covar)) {
-    between <- between[!is_covar]
-    warning("Levene's test is not appropriate with quantitative explanatory variables. ",
-            "Testing assumption of homogeneity among factor groups only.")
-  }
-
-  p.val <- test[1, "Pr(>F)"]
-
-  if (p.val < 0.05) {
-    insight::print_color(sprintf("Warning: Heteroscedasticity (non-constant error variance) detected (%s).\n", insight::format_p(p.val)), "red")
-  } else {
-    insight::print_color(sprintf("OK: Error variance appears to be homoscedastic (%s).\n", insight::format_p(p.val)), "green")
-  }
-
-  attr(p.val, "object_name") <- deparse(substitute(x), width.cutoff = 500)
-  class(p.val) <- unique(c("check_heteroscedasticity", "see_check_heteroscedasticity", class(p.val)))
-
-  invisible(p.val)
-
 }
 
 
