+++ man/check_heteroscedasticity.Rd
@@ -23,8 +23,7 @@ Check model for (non-)constant error variance.
 }
 \details{
 This test of the hypothesis of (non-)constant error is also called
-  \emph{Breusch-Pagan test} (\cite{1979}). For \code{afex_aov} Levene's test
-  is preformed.
+  \emph{Breusch-Pagan test} (\cite{1979}).
 }
 \note{
 There is also a \href{https://easystats.github.io/see/articles/performance.html}{\code{plot()}-method} implemented in the \href{https://easystats.github.io/see/}{\pkg{see}-package}.
