+++ NAMESPACE
@@ -25,8 +25,8 @@ S3method(check_convergence,default)
 S3method(check_convergence,merMod)
 S3method(check_distribution,default)
 S3method(check_distribution,numeric)
-S3method(check_heteroscedasticity,afex_aov)
 S3method(check_heteroscedasticity,default)
+S3method(check_homogeneity,afex_aov)
 S3method(check_homogeneity,default)
 S3method(check_model,default)
 S3method(check_model,model_fit)
