+++ DESCRIPTION
@@ -7,22 +7,22 @@ Authors@R:
              family = "Lüdecke",
              role = c("aut", "cre"),
              email = "d.luedecke@uke.de",
-             comment = c(ORCID = "0000-0002-8895-3206")),
+             comment = c(ORCID = "0000-0002-8895-3206", Twitter = "@strengejacke")),
       person(given = "Dominique",
              family = "Makowski",
              role = c("aut", "ctb"),
              email = "dom.makowski@gmail.com",
-             comment = c(ORCID = "0000-0001-5375-9967")),
+             comment = c(ORCID = "0000-0001-5375-9967", Twitter = "@Dom_Makowski")),
      person(given = "Mattan S.",
              family = "Ben-Shachar",
              role = c("aut", "ctb"),
              email = "matanshm@post.bgu.ac.il",
-             comment = c(ORCID = "0000-0002-4287-4801")),
+             comment = c(ORCID = "0000-0002-4287-4801", Twitter = "@mattansb")),
       person(given = "Indrajeet",
              family = "Patil",
              role = c("aut", "ctb"),
              email = "patilindrajeet.science@gmail.com",
-             comment = c(ORCID = "0000-0003-1995-6531")),
+             comment = c(ORCID = "0000-0003-1995-6531", Twitter = "@patilindrajeets")),
       person(given = "Philip",
              family = "Waggoner",
              role = c("aut", "ctb"),
