+++ R/check_distribution.R
@@ -121,7 +121,7 @@ check_distribution.default <- function(model) {
   data.frame(
     "SD" = stats::sd(x),
     "MAD" = stats::mad(x, constant = 1),
-    "Mean_Median_Distance" = mean(x) - median(x),
+    "Mean_Median_Distance" = mean(x) - stats::median(x),
     "Mean_Mode_Distance" = mean(x) - as.numeric(bayestestR::map_estimate(x, bw = "nrd0")),
     "SD_MAD_Distance" = stats::sd(x) - stats::mad(x, constant = 1),
     "Var_Mean_Distance" = stats::var(x) - mean(x),
