### 2021-04-18 4c5990c
Update r2_ci.R
