+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: 8e0838e273462cec420dfa20f240c684a33425d9
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-17T22:24Z
+last_built: 2021-04-18T09:03Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
