### 2021-04-19 9d177a4
use model.matrix for paper
