### 2021-04-19 #3139 @strengejacke
@mikldk

- I have done final proof reading. Since we have a native speaker as co-author, I assume that language and style are ok.
- I have checked the generated wheedon PDF proof
- I created a release on GitHub for this version that is in line with the version related to the paper
- I have created a Zenodo-archive of this release and edited/checked the meta-data.

I confirm that the GitHub release as well as the Zenodo-archive are the correct package versions and in the state that should be published.

The doi is `10.5281/zenodo.4700887`.
