+++ R/test_bf.R
@@ -5,8 +5,9 @@ test_bf <- function(...) {
 }
 
 
+#' @rdname test_performance
 #' @export
-test_bf.default <- function(...) {
+test_bf.default <- function(..., text_length = NULL) {
 
   # Attribute class to list and get names from the global environment
   objects <- insight::ellipsis_info(..., only_models = TRUE)
@@ -21,7 +22,7 @@ test_bf.default <- function(...) {
 
   # If a suitable class is found, run the more specific method on it
   if (inherits(objects, c("ListNestedRegressions", "ListNonNestedRegressions", "ListLavaan"))) {
-    test_bf(objects)
+    test_bf(objects, text_length = text_length)
   } else {
     stop("The models cannot be compared for some reason :/")
   }
@@ -30,7 +31,7 @@ test_bf.default <- function(...) {
 
 
 #' @export
-test_bf.ListModels <- function(objects, reference = 1, ...) {
+test_bf.ListModels <- function(objects, reference = 1, text_length = NULL, ...) {
   if (.test_bf_areAllBayesian(objects) == "mixed") {
     stop("You cannot mix Bayesian and non-Bayesian models in 'test_bf()'.", call. = FALSE)
   }
@@ -62,6 +63,7 @@ test_bf.ListModels <- function(objects, reference = 1, ...) {
 
   # Replace denominator
   attr(rez, "denominator") <- ref
+  attr(rez, "text_length") <- text_length
   class(rez) <- c("bayesfactor_models", "see_bayesfactor_models", class(rez))
   rez
 }
