+++ DESCRIPTION
@@ -105,4 +105,5 @@ Suggests:
 Encoding: UTF-8
 Language: en-US
 RoxygenNote: 7.1.1
+Remotes: easystats/insight
 Config/testthat/edition: 3
