+++ NAMESPACE
@@ -528,6 +528,7 @@ importFrom(stats,bartlett.test)
 importFrom(stats,binomial)
 importFrom(stats,coef)
 importFrom(stats,complete.cases)
+importFrom(stats,contr.sum)
 importFrom(stats,cooks.distance)
 importFrom(stats,cor)
 importFrom(stats,cov)
