+++ R/check_collinearity.R
@@ -100,8 +100,6 @@ check_collinearity.default <- function(x, verbose = TRUE, ...) {
 }
 
 #' @export
-#' @importFrom stats contr.sum lm
-#' @importFrom insight find_response get_data
 check_collinearity.afex_aov <- function(x, verbose = TRUE, ...){
   f <- paste(row.names(x$anova_table), collapse = "+")
   f <- paste0(insight::find_response(x), "~", f)
