+++ NAMESPACE
@@ -484,67 +484,3 @@ export(variance_decomposition)
 importFrom(insight,display)
 importFrom(insight,print_html)
 importFrom(insight,print_md)
-importFrom(stats,AIC)
-importFrom(stats,BIC)
-importFrom(stats,IQR)
-importFrom(stats,anova)
-importFrom(stats,as.formula)
-importFrom(stats,bartlett.test)
-importFrom(stats,binomial)
-importFrom(stats,coef)
-importFrom(stats,complete.cases)
-importFrom(stats,contr.sum)
-importFrom(stats,cooks.distance)
-importFrom(stats,cor)
-importFrom(stats,cov)
-importFrom(stats,cov2cor)
-importFrom(stats,dbinom)
-importFrom(stats,density)
-importFrom(stats,df.residual)
-importFrom(stats,dnbinom)
-importFrom(stats,dpois)
-importFrom(stats,family)
-importFrom(stats,fitted)
-importFrom(stats,fligner.test)
-importFrom(stats,formula)
-importFrom(stats,glm)
-importFrom(stats,influence)
-importFrom(stats,lm)
-importFrom(stats,mad)
-importFrom(stats,mahalanobis)
-importFrom(stats,median)
-importFrom(stats,model.frame)
-importFrom(stats,model.matrix)
-importFrom(stats,na.omit)
-importFrom(stats,nobs)
-importFrom(stats,pchisq)
-importFrom(stats,pf)
-importFrom(stats,plogis)
-importFrom(stats,pnbinom)
-importFrom(stats,pnorm)
-importFrom(stats,ppoints)
-importFrom(stats,ppois)
-importFrom(stats,predict)
-importFrom(stats,predict.glm)
-importFrom(stats,qchisq)
-importFrom(stats,qf)
-importFrom(stats,qnorm)
-importFrom(stats,quantile)
-importFrom(stats,reshape)
-importFrom(stats,residuals)
-importFrom(stats,rstandard)
-importFrom(stats,rstudent)
-importFrom(stats,sd)
-importFrom(stats,setNames)
-importFrom(stats,shapiro.test)
-importFrom(stats,simulate)
-importFrom(stats,summary.lm)
-importFrom(stats,terms)
-importFrom(stats,update)
-importFrom(stats,var)
-importFrom(stats,vcov)
-importFrom(stats,weights)
-importFrom(stats,xtabs)
-importFrom(utils,capture.output)
-importFrom(utils,packageVersion)
-
