### 2021-04-20 32119d6
exclude multinominal & polr
