+++ R/r2_mcfadden.R
@@ -60,7 +60,7 @@ r2_mcfadden <- function(model, ...) {
 #' @export
 r2_mcfadden.glm <- function(model, verbose = TRUE, ...) {
   info <- insight::model_info(model)
-  if (info$is_binomial && !info$is_bernoulli) {
+  if (info$is_binomial && !info$is_bernoulli && class(model)[1] == "glm") {
     if (verbose) {
       warning("Can't calculate accurate R2 for binomial models\n  that are not Bernoulli models.", call. = FALSE)
     }
