### 2021-04-21 ce67afa
add: tests for afex_aov
