### 2021-04-21 0c0077d
show afex in check_homogeneity docs
