### 2021-04-21 1a8dbbd
fix: check_homogeneity
