+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: 46764c0dba9c847957120d0bd687862346db9e98
 articles:
   compare: compare.html
   r2: r2.html
-last_built: 2021-04-19T22:00Z
+last_built: 2021-04-21T06:16Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
